infra/modules/alerting/function/requirements.txt
functions-framework
google-cloud-secret-manager
requests
